Update color palette and scatterplot radius defaults
