+++ src/layers/index.js
@@ -1,2 +1,3 @@
 export { SelectablePolygonLayer } from './SelectablePolygonLayer';
 export { SelectableScatterplotLayer } from './SelectableScatterplotLayer';
+export { getSelectionLayers } from './utils';
